// frontend/src/types/asset-form-values.ts
export interface AssetFormValues {
  readonly name: string
  readonly sectorName: string
  readonly ipAddress: string
  readonly macAddress: string
  readonly wifiMacAddress: string
  readonly modelName: string
  readonly userName: string
  readonly processor: string
  readonly memory: string
  readonly storage: string
  readonly brand: string
  readonly partNumber: string
  readonly patrimony: string
  readonly desktopName: string
  readonly assetType: string
  readonly operatingSystem: string
}